pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.29.0",
    "llama-index>=0.13.1",
    "llama-index-core>=0.13.1",
    "llama-index-embeddings-openai>=0.5.0",
    "llama-index-llms-openai>=0.5.2",
    "openai>=1.99.7",
    "pandas>=2.2.3",
    "streamlit>=1.48.0",
]
